@@ -4,34 +4,53 @@
     "@emotion/styled": "^11.9.3",
     "@mui/material": "^5.8.4",
     "content-scripts-register-polyfill": "^3.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "regenerator-runtime": "^0.13.9",
     "styled-components": "^5.3.5",
     "workboots": "^1.0.3"
   },
   "scripts": {
     "dev": "NODE_ENV=development webpack --mode development --config webpack.develop.js --watch",
-    "production": "NODE_ENV=development webpack --mode development --config webpack.prod.js --watch"
+    "production": "NODE_ENV=development webpack --mode development --config webpack.prod.js --watch",
+    "test": "npx jest"
   },
   "devDependencies": {
     "@babel/core": "^7.17.5",
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-react": "^7.16.7",
     "@babel/preset-typescript": "^7.16.7",
+    "@testing-library/dom": "^9.3.1",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "@types/chrome": "^0.0.180",
+    "@types/jest": "^29.5.3",
     "@types/react": "^17.0.43",
     "@types/react-dom": "^17.0.14",
+    "@types/react-test-renderer": "^18.0.0",
     "@types/styled-components": "^5.1.24",
+    "@types/testing-library__react": "^10.2.0",
+    "@typescript-eslint/eslint-plugin": "^6.2.0",
+    "@typescript-eslint/parser": "^6.2.0",
     "babel-jest": "^27.5.1",
     "babel-loader": "^8.2.3",
-    "copy-webpack-plugin": "^10.2.4",
-    "css-loader": "^6.6.0",
+    "caniuse-lite": "^1.0.30001517",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.8.1",
+    "eslint": "^8.45.0",
+    "eslint-webpack-plugin": "^4.0.1",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.0",
-    "jest": "^27.5.1",
+    "fork-ts-checker-webpack-plugin": "^8.0.0",
+    "html-webpack-plugin": "^5.5.3",
+    "jest": "^29.6.2",
+    "jest-environment-jsdom": "^29.6.2",
+    "mini-css-extract-plugin": "^2.7.6",
+    "node-sass": "^9.0.0",
     "react-test-renderer": "^18.0.0",
-    "style-loader": "^3.3.1",
+    "sass-loader": "^13.3.2",
+    "style-loader": "^3.3.3",
+    "ts-jest": "^29.1.1",
     "ts-loader": "^9.2.8",
     "typescript": "^4.6.3",
     "webpack": "^5.69.1",
