@@ -1,6 +1,6 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import App from "../src/app.tsx";
 import "./index.css";
 
-ReactDOM.render(<App />, document.querySelector("#root"));
+ReactDOM.createRoot(document.querySelector("#root")).render(<App />);
